2.9 auth

+++ index.js
@@ -39,20 +39,7 @@ app.get('/documentation', (req, res) => {
     res.sendFile('public/documentation.html', { root: __dirname });
 });
 
-// Gets the list of data about all movies
-// app.get('/movies', passport.authenticate('jwt', {session: false} ), async (req, res) => {
-//     await Movies.find()
-//     .then((movies) => {
-//         res.status(201).json(movies);
-//     })
-//     .catch((err) => {
-//         console.error(err);
-//         res.status(500).send("Error: " + err);
-//     })
-// });
-
-// non-redundant version
-app.get('/movies', passport.authenticate('jwt', {session: false} ), (req, res) => {
+app.get('/movies', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.find()
     .then((movies) => {
         res.status(201).json(movies);
@@ -63,19 +50,7 @@ app.get('/movies', passport.authenticate('jwt', {session: false} ), (req, res) =
     })
 });
 
-// Gets the data about a single movie (by title)
-// app.get('/movies/:title', async (req, res) => {
-//     await Movies.findOne({Title: req.params.title})
-//     .then((movies) => {
-//         res.status(201).json(movies);
-//     })
-//     .catch((err) => {
-//         console.error(err);
-//         res.status(500).send("Error: " + err);
-//     })
-// });
-//non-redundant version
-app.get('/movies/:title', (req, res) => {
+app.get('/movies/:title', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.findOne({Title: req.params.title})
     .then((movies) => {
         res.status(201).json(movies);
@@ -87,7 +62,7 @@ app.get('/movies/:title', (req, res) => {
 });
 
 // Gets the data about a genre by name of genre
-app.get('/movies/genres/:genre', (req, res) => {
+app.get('/movies/genres/:genre', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.findOne({"Genre.Name": req.params.genre})
     .then((movie) => {
         res.status(201).json(movie.Genre);
@@ -99,7 +74,7 @@ app.get('/movies/genres/:genre', (req, res) => {
 });
 
 // Get data about a director (bio, birth year, death year) by name;
-app.get('/movies/directors/:director', (req, res) => {
+app.get('/movies/directors/:director', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.findOne({"Director.Name": req.params.director})
     .then((movie) => {
         res.status(201).json(movie.Director);
@@ -111,7 +86,7 @@ app.get('/movies/directors/:director', (req, res) => {
 });
 
 // Get all users
-app.get('/users', (req, res) => {
+app.get('/users', passport.authenticate('jwt', {session: false}), (req, res) => {
     Users.find()
         .then((users) => {
             res.status(201).json(users);
@@ -123,7 +98,7 @@ app.get('/users', (req, res) => {
 });
 
 // Get specific user
-app.get('/users/:Username', (req, res) => {
+app.get('/users/:Username', passport.authenticate('jwt', {session: false}), (req, res) => {
     Users.findOne({ Username: req.params.Username })
         .then((user) => {
             res.json(user);
@@ -143,10 +118,10 @@ app.get('/users/:Username', (req, res) => {
   Email: String,
   BirthDate: Date
 }*/
-
+// Note: Do not authenticate here, obviously
 // Allow new users to register
-app.post('/users', async (req, res) => {
-    await Users.findOne({ Username: req.body.Username })
+app.post('/users', (req, res) => {
+    Users.findOne({ Username: req.body.Username })
         .then((user) => {
             if (user) {
                 return res.status(400).send(req.body.Username + 'already exists');
@@ -182,7 +157,10 @@ app.post('/users', async (req, res) => {
   (required)
   BirthDate: Date
 }*/
-app.put('/users/:Username', (req, res) => {
+app.put('/users/:Username', passport.authenticate('jwt', {session: false}), (req, res) => {
+    if (req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
     Users.findOneAndUpdate({ Username: req.params.Username },
         {
             $set:
@@ -205,7 +183,10 @@ app.put('/users/:Username', (req, res) => {
 
 
 // Allow users to add a movie to their list of favorites (showing only a text that a movie has been added - more on this later);
-app.put('/users/:Username/movies/:MovieID', (req, res) => {
+app.put('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), (req, res) => {
+    if (req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
     Users.findOneAndUpdate({ Username: req.params.Username },
         {
             // note to self: consider replacing $push with $addToSet to avoid duplicates
@@ -222,7 +203,10 @@ app.put('/users/:Username/movies/:MovieID', (req, res) => {
 });
 
 // Allow users to remove a movie from their list of favorites (showing only a text that a movie has been removed - more on this later);
-app.delete('/users/:Username/movies/:MovieID', (req, res) => {
+app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), (req, res) => {
+    if (req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
     Users.findOneAndUpdate({ Username: req.params.Username },
         {
             $pull: { FavoriteMovies: req.params.MovieID }
@@ -238,7 +222,10 @@ app.delete('/users/:Username/movies/:MovieID', (req, res) => {
 });
 
 // Allow existing users to deregister (showing only a text that a user email has been removed - more on this later).
-app.delete('/users/:Username', (req, res) => {
+app.delete('/users/:Username', passport.authenticate('jwt', {session: false}), (req, res) => {
+    if (req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
     Users.findOneAndDelete({ Username: req.params.Username })
         .then((user) => {
             if (!user) {
